updating template and Charm-C++ group API

+++ auto_outsrc/templates/build_miracl.sh
@@ -13,6 +13,11 @@ if [ -z $curve ]; then
    echo "Building default curve in miracl: $curve"
 fi
 
+if [ -e miracl-$curve.a ]; then
+    echo "Already built miracl-$curve" 
+    exit 0
+fi
+
 cp mirdef.hpp mirdef.h
 g++ -c -m64 -O2 mrcore.c
 g++ -c -m64 -O2 mrarth0.c
@@ -75,9 +80,9 @@ g++ -c -m64 -O2 ec2.cpp
 g++ -c -m64 -O2 flash.cpp
 g++ -c -m64 -O2 crt.cpp
 # Cocks-Pinch curve
-g++ -c -m64 -O2 cp_pair.cpp
+# g++ -c -m64 -O2 cp_pair.cpp
 # Barreto-Naehrig curve
-g++ -c -m64 -O2 bn_pair.cpp
+# g++ -c -m64 -O2 bn_pair.cpp
 ar rc miracl.a mrcore.o mrarth0.o mrarth1.o mrarth2.o mralloc.o mrsmall.o mrzzn2.o mrzzn3.o
 ar r miracl.a mrio1.o mrio2.o mrjack.o mrgcd.o mrxgcd.o mrarth3.o mrbits.o mrecn2.o
 ar r miracl.a mrrand.o mrprime.o mrcrt.o mrscrt.o mrmonty.o mrcurve.o mrsroot.o mrzzn2b.o
@@ -86,11 +91,13 @@ ar r miracl.a mrflash.o mrfrnd.o mrdouble.o mrround.o mrbuild.o
 ar r miracl.a mrflsh1.o mrpi.o mrflsh2.o mrflsh3.o mrflsh4.o 
 ar r miracl.a mrbrick.o mrebrick.o mrec2m.o mrgf2m.o mrmuldv.o mrshs512.o
 #ar r miracl.a big.o zzn.o zzn2.o zzn3.o zzn6a.o ecn.o ecn3.o ec2.o flash.o crt.o mnt_pair.o 
-if [ $curve = 'kss' ]; then
-   # KSS curve
-   g++ -c -m64 -O2 kss_pair.cpp zzn18.cpp zzn6.cpp ecn3.cpp zzn3.cpp
+if [ $curve = 'ssp' ]; then
+   # SSP curve
+   #g++ -c -m64 -O2 kss_pair.cpp zzn18.cpp zzn6.cpp ecn3.cpp zzn3.cpp
+   g++ -c -m64 -O2 ssp_pair.cpp zzn2.cpp
    cp miracl.a miracl-$curve.a
-   ar r miracl-$curve.a big.o zzn.o zzn3.o zzn6.o zzn18.o ecn.o ecn3.o ec2.o flash.o crt.o kss_pair.o
+   #ar r miracl-$curve.a big.o zzn.o zzn3.o zzn6.o zzn18.o ecn.o ecn3.o ec2.o flash.o crt.o kss_pair.o
+   ar r miracl-$curve.a big.o zzn.o zzn2.o ecn.o ec2.o flash.o crt.o ssp_pair.o
 elif [ $curve = 'mnt' ]; then
    # MNT curve
    g++ -c -m64 -O2 mnt_pair.cpp zzn6a.cpp ecn3.cpp zzn3.cpp zzn2.cpp
